seamless-mode: fix windows leaving traces
In real-world conditions, messages with seamless changes might get delayed
in comparison to display data. While the client has outdated positions of
windows, spice_cairo_draw_event() draws parts of the display that should not
be visible any more at the moment. Once the data with new positions arrive,
there might be no need to redraw the old position because it has been redrawn
earlier and hence the cairo is clipped in such a way that does not allow
to draw the appropriate areas transparent.

This issue can be reproduced by moving/resizing windows very quickly or
by manually delaying window changes events on the agent side.

Opposite issue happens when a window is brought from minimalized state.
Outdated seamless-mode list prevents the window from drawing properly.
The border of the window is drawn first, the inner (client) area later.
If the client receives updated coordinates after the border is drawn and
cairo is then clipped on the client area, the border stays transparent.

Becuase of these reasons, the old and new seamless mode region should be
invalidated after we receive new seamless mode list/change.