main: support VD_AGENT_SELECTION_* messages
These enable transferring arbitrary type of data
unlike the VDAgentClipboard* messages that are
restricted to types defined in spice protocol
(and atom2agent[] in spice-gtk-session.c).

Messages received from the agent are checked for malformations
and corresponding signals ("main-selection-*") are subsequently
emitted on the SpiceMainChannel object.

Add spice_main_channel_selection_*() functions to send
VD_AGENT_SELECTION_* messsages to the vdagent.
It is an error to call any of these functions
if the connected vdagent does not feature
VD_AGENT_CAP_SELECTION_DATA capability.

This will be used later for clipboard and DND.

Signed-off-by: Jakub Janků <jjanku@redhat.com>